remove forEachComponent()
